+++ config/babelTransform.js
@@ -4,5 +4,5 @@ const babelJest = require('babel-jest')
 const createBabelConfig = require('./createBabelConfig')
 
 module.exports = babelJest.createTransformer(
-  createBabelConfig({ modules: 'commonjs' })
+  createBabelConfig({ modules: 'commonjs', isTestEnvironment: true })
 )